fix broken script
The frontier forum has changed since last visit months ago, with new URL
scheme used and SSL only. This required both a script update as well as
a backend update.

* Updates forum URL parsing of the current page.
* Reduces privileges the script requires, by having reconfigured the
  backend to allow plain browser XHR using cross origin directives.
* Switched backend to SSL, since the forum is now SSL only (browsers
  don't allow downgrades, sourcing unencrypted content from encrypted
  pages.

Since the Greasmonkey GM-XHR function is no longer used, the page can
run in any fashion on a web page, be it simply injected using the
browser's dev tool, as a GM script or as an extension (Firefox jetpack
or Google Chrome extension).