Speed up browser rendering
Only the last 100 dev posts are rendered by default.
A control has been added right of the "DEV POSTS" button, which allows
users to toggle showing all posts, as in previous versions.

This change is introduced because of user feedback on the forum:
https://forums.frontier.co.uk/showthread.php/40194-Quickly-find-all-Frontier-Developer-posts/page3

As there are now more than 4000 dev posts, even the fastest browser on
the fastest processor is struggling to render all that data, so it was
time to do _something_ about the bad user experience.